REDDIT/requirements.txt
flask
gunicorn
requests
python-dotenv==0.21.0
